escape the item["id"]
this is now necessary because we added redirects
to every single disambiguation page with unbalanced
parentheses in order to make messed-up links from
reddit and other sites using markdown not break
in the case that the user posting the link neglects
to escape the trailing closed parenthesis.

For example, Farmer (George Normore now redirects to
Farmer (George Normore).

Because we were missing the re.escape previously in
this line of code, the presence of this redirect on
the wiki now created a syntax error in every single
scoreboard parse.