Split up lookup cache queries into two parts
Module:Teamnames got too big and the expandtemplates
api limit could no longer process queries of JsonEncode
expanding this module for the lookup_cache to parse it.

Module:JsonEncode has now been updated to accept two
parameters, `include_match` and `exclude_match`, and
we are now taking advantage of these two parameters
to split up the query into to parts so that we don't
reach the api limit (or even come close to it) anymore.